created functional new edit page and route
